backend/requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
# Add gunicorn or waitress for production deployment